Refactor ファイルの格納ディレクトリの移動
