+++ src/view/plandetail/CollageTemplate.tsx
